# packages/queries/output.swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TibArena API",
    "description": "The TibArena API documentation",
    "version": "1.0",
    "contact": {}
  },
  "servers": [],
  "paths": {
...
    "/api/example/protected": {
      "get": {
        "operationId": "ExampleController_getProtectedData",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successfully retrieved protected data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "You have access to this protected resource"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "example": "1234567890"
                        },
                        "walletAddress": {
                          "type": "string",
                          "example": "0x1234567890abcdef1234567890abcdef12345678"
                        },
                        "username": {
                          "type": "string",
                          "example": "user123"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing authentication"
          }
        },
        "summary": "Get protected user data",
        "tags": [
          "Authentication Example"
        ]
      }
    },
    "/api/jwt-example/protected": {
      "get": {
        "operationId": "SimplifiedExampleController_getProtectedData",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successfully retrieved protected data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "You have access to this JWT-protected resource"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "example": "1234567890"
                        },
                        "walletAddress": {
                          "type": "string",
                          "example": "0x1234567890abcdef1234567890abcdef12345678"
                        },
                        "username": {
                          "type": "string",
                          "example": "user123"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing JWT token"
          }
        },
        "summary": "Get protected user data using JWT",
        "tags": [
          "JWT Authentication Example"
        ]
      }
    }
  },
  "components": {
...
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken"
      }
    }
  }
}